chore: merge editor-dev to dev (#403)
* feat(rich-text-editor): add new component rich-text-editor (#346)

* feat(rich-text-editor): add new components rich-text-editor

---------


* style(rich-text-editor): adjust style, fit Highlight, Default, Disabled (#357)

* style(rich-text-editor): adjust style, fit Highlight, Default, Disabled

* fix(rich-text-editor): Space format changed from four to two

* fix(rich-text-editor): add svg file and support high light and disabled and default

---------


* feat(rich-text-editor): add text alignment, task list, table interaction (#372)

* feat(rich-text-editor): add text alignment, task list, table interaction

---------


* perf(rich-text-editor): split pure function from vue.ts to index.ts (#395)


* fix(rich-text-editor): fix conflict

---------
