+++ packages/theme/src/rich-text-editor/vars.less
@@ -1,1 +1,9 @@
-.component-css-vars-roles() {}
+.component-css-vars-rich-text-editor() {
+  --ti-rich-text-editor-box-outline-color: var(--ti-common-color-line-normal);
+  --ti-rich-text-editor-box-outline-hover-color: var(--ti-common-color-line-hover);
+  --ti-rich-text-editor-button-hover: var(--ti-base-color-brand-8);
+  --ti-rich-text-editor-button-disabled: var(--ti-common-color-primary-disabled-bgcolor);
+  --ti-rich-text-editor-button-active: var(--ti-base-color-brand-8);
+  --ti-rich-text-editor-poplist-item-selected-bg-color: var(--ti-common-color-selected-background);
+  --ti-rich-text-editor-poplist-item-selected-text-color: var(--ti-common-color-selected-text-color);
+}
